requirements.txt
amqp==2.6.1
appdirs==1.4.3
asgiref==3.2.10
astroid==2.4.2
billiard==3.6.3.0
CacheControl==0.12.6
celery==4.4.7
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
defusedxml==0.6.0
distlib==0.3.0
distro==1.4.0
dj-database-url==0.5.0
Django==3.1
django-allauth==0.42.0
django-cors-headers==3.5.0
django-heroku==0.3.1
django-rest-auth==0.9.5
djangorestframework==3.11.1
gunicorn==20.0.4
html5lib==1.0.1
idna==2.10
install-requires==0.3.0
ipaddr==2.2.0
isort==4.3.21
kombu==4.6.11
lazy-object-proxy==1.4.3
lockfile==0.12.2
mccabe==0.6.1
msgpack==0.6.2
oauthlib==3.1.0
packaging==20.3
pep517==0.8.2
Pillow==7.2.0
setuptools==50.3.2
progress==1.5
psycopg2==2.8.5
pycparser==2.20
pydotenv==0.0.7
PyJWT==1.7.1
pylint==2.5.3
pylint-django==2.3.0
pylint-plugin-utils==0.6
pyparsing==2.4.6
python-decouple==3.3
python-dotenv==0.14.0
python3-openid==3.2.0
pytoml==0.1.21
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
retrying==1.3.3
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.10
vine==1.3.0
webencodings==0.5.1
whitenoise==5.2.0
wrapt==1.12.1
zipp==3.1.0

